+++ api/internal/plugins/loader/loader.go
@@ -168,9 +168,9 @@ func (l *Loader) makeBuiltinPlugin(r resid.Gvk) (resmap.Configurable, error) {
 }
 
 func (l *Loader) loadPlugin(res *resource.Resource) (resmap.Configurable, error) {
-	_, err := execplugin.GetFunctionSpec(res)
-	if err == nil {
-		return execplugin.NewFnPlugin(&l.pc.FnpLoadingOptions), nil
+	spec := fnplugin.GetFunctionSpec(res)
+	if spec != nil {
+		return fnplugin.NewFnPlugin(&l.pc.FnpLoadingOptions), nil
 	}
 	return l.loadExecOrGoPlugin(res.OrgId())
 }
